refactor: independent settings folder, env
