+++ backend/naengpa/naengpa/asgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'naengpa.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'settings.local')
 
 application = get_asgi_application()
