+++ backend/naengpa/settings/base.py
@@ -1,47 +1,18 @@
 """
 Django settings for naengpa project.
-
-Generated by 'django-admin startproject' using Django 3.0.8.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.0/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.0/ref/settings/
 """
-
 import os
-from dotenv import load_dotenv
-from django.core.exceptions import ImproperlyConfigured
-
-load_dotenv()
-
-
-def get_env_value(var_name):
-    """help function for os.getenv to handle KeyError"""
-    try:
-        return os.getenv(var_name)
-    except KeyError:
-        error_msg = 'Set the {} environment variable'.format(var_name)
-        raise ImproperlyConfigured(error_msg) from None
+from .env import get_env_value
 
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/3.0/howto/deployment/checklist/
-
-# SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = get_env_value('SECRET_KEY')
 
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = False
+DEBUG = True  # to be overridden
 
-
-AWS_EC2_ELASTIC_IP = get_env_value('AWS_EC2_ELASTIC_IP')
-ALLOWED_HOSTS = [AWS_EC2_ELASTIC_IP]
+ALLOWED_HOSTS = []  # to be overridden
 
 
 # Application definition
@@ -99,8 +70,6 @@ WSGI_APPLICATION = 'naengpa.wsgi.application'
 
 
 # Database
-# https://docs.djangoproject.com/en/3.0/ref/settings/#databases
-
 DATABASES = {
     'default': {
         'ENGINE': 'django.contrib.gis.db.backends.postgis',
@@ -157,29 +126,10 @@ AUTH_PASSWORD_VALIDATORS = [
 # https://docs.djangoproject.com/en/3.0/topics/i18n/
 
 LANGUAGE_CODE = 'en-us'
-
 TIME_ZONE = 'Asia/Seoul'
-
 USE_I18N = True
-
 USE_L10N = True
-
 USE_TZ = True
 
-
 # Static files (CSS, JavaScript, Images)
-# https://docs.djangoproject.com/en/3.0/howto/static-files/
-
 STATIC_URL = '/static/'
-
-# AWS S3 BUCKET
-S3_URL = get_env_value("S3_URL")
-DEFAULT_FILE_STORAGE = get_env_value("DEFAULT_FILE_STORAGE")
-AWS_ACCESS_KEY_ID = get_env_value("AWS_ACCESS_KEY_ID")
-AWS_SECRET_ACCESS_KEY = get_env_value("AWS_SECRET_ACCESS_KEY")
-AWS_STORAGE_BUCKET_NAME = get_env_value("AWS_STORAGE_BUCKET_NAME")
-AWS_S3_REGION_NAME = get_env_value("AWS_S3_REGION_NAME")
-AWS_S3_FILE_OVERWRITE = get_env_value("AWS_S3_REGION_NAME")
-
-# Kakao REST Api key
-APP_REST_API_KEY = get_env_value("APP_REST_API_KEY")
