+++ backend/naengpa/settings/local.py
@@ -1,0 +1,7 @@
+# pylint: skip-file
+from .base import *
+
+
+DEBUG = True
+
+ALLOWED_HOSTS = ['localhost']
